+++ docs/source/changelog.rst
@@ -1,6 +1,14 @@
 Changelog
 =========
 
+v1.4.1 (2016-??-??)
+-------------------
+
+#. Added ``DFXP`` sync map output format, alias for ``TTML``
+#. Added ``SBV`` sync map output format (SubViewer format with newline, used by YouTube)
+#. Added ``SUB`` sync map output format (SubViewer format with ``[br]``)
+#. Marked Czech (``cs``) language as tested
+
 v1.4.0 (2016-01-15)
 -------------------